feat(institutions): refactor en estructura frontend
