+++ apps/frontend/src/app/institutions/institution.routing.ts
@@ -1,8 +1,8 @@
 import { Routes, RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 
-import { AppInstitutionCrudComponent } from './institution-crud.component';
-import { AppInstitutionListComponent } from './institution-list.component';
+import { AppInstitutionCrudComponent } from './components/institutions-crud/institution-crud.component';
+import { AppInstitutionListComponent } from './components/institutions-list/institution-list.component';
 
 const routes: Routes = [
     {
